+++ packages/algolia-search/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @graphcommerce/algolia-search
 
+## 9.0.2
+
+## 9.0.2-canary.0
+
 ## 9.0.1
 
 ## 9.0.1-canary.1