+++ packages/demo-magento-graphcommerce/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @graphcommerce/demo-magento-graphcommerce
 
+## 9.0.2
+
+## 9.0.2-canary.0
+
 ## 9.0.1
 
 ## 9.0.1-canary.1