+++ packages/magento-product/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 9.0.2
+
+## 9.0.2-canary.0
+
 ## 9.0.1
 
 ### Patch Changes